[xinit] compositing
Former-commit-id: 4b03b23f9898f75731a0691b3d7d51bc951e0e2a [formerly 96814e30d2a4ac0cb8265bdfb13eaa26960a1ea5] [formerly 9295dcd00eaa248138f123147be0d23e2fa9ebe5 [formerly ef50aaf544229766f2e6dd481d5ec9fd83dcbca5]]
Former-commit-id: 547c2f46818fb9dac84d943e9764b0db3b29df2d [formerly 5c9219fcb5437bd5446908eab76f1ac5d9a5344b]
Former-commit-id: 5bdf284c49e14fb7d81ac8d84be5ce5cacb2d250

+++ .config/xinit/compositing.zsh
@@ -1,4 +1,6 @@
 #!/bin/zsh
 
-nexec compton || sudo pkill compton
-compton -b --config ${XDG_CONFIG_HOME}/compton/compton.conf &
+if [[ -z "$1" ]]; then
+    nexec compton || sudo pkill compton
+    compton -b --config ${XDG_CONFIG_HOME}/compton/compton.conf &
+fi
